Structure separated mesh even further
The previous structure had two problems:
 1. There were several circular dependencies between the parts that were
    hard to resolve
 2. It was not clear which file to include, and each of them would have
    introduced a different include ordering.

This commit introduces a top level Mesh.h and moves all details into the
MeshImpl folder. Doings so we get a single header to include for the
Mesh with all its parts. Moreover, this single file is tasked with
assembling everything into a complete header. The individual header
files don't have to include anything reducing all the include ordering
mess.
